refactor : UploadFileInfoDTO 리팩토링
- 해당 DTO의 위치를 content 도메인으로 옮겼습니다.
- DTO를 record로 변경하였습니다.
