+++ src/main/java/com/example/projectlxp/content/service/dto/UploadFileInfoDTO.java
@@ -1,0 +1,5 @@
+package com.example.projectlxp.content.service.dto;
+
+import com.example.projectlxp.lecture.entity.LectureType;
+
+public record UploadFileInfoDTO(String fileURL, LectureType fileType, String videoDuration) {}
