--- src/main/java/com/example/projectlxp/lecture/controller/dto/UploadFileInfoDTO.java
@@ -1,17 +1,0 @@
-package com.example.projectlxp.lecture.controller.dto;
-
-import com.example.projectlxp.lecture.entity.LectureType;
-
-public class UploadFileInfoDTO {
-    public String fileURL;
-
-    public LectureType fileType;
-
-    public String videoDuration;
-
-    public UploadFileInfoDTO(String fileURL, LectureType fileExt, String videoDuration) {
-        this.fileURL = fileURL;
-        this.fileType = fileExt;
-        this.videoDuration = videoDuration;
-    }
-}
